dispatch schemas by kind
