+++ main/src/ca/uwaterloo/flix/language/phase/unification/Unification.scala
@@ -75,6 +75,8 @@ object Unification {
 
       case _ if tpe1.kind == Kind.RecordRow && tpe2.kind == Kind.RecordRow => RecordUnification.unifyRows(tpe1, tpe2, renv)
 
+      case _ if tpe1.kind == Kind.SchemaRow && tpe2.kind == Kind.SchemaRow => SchemaUnification.unifyRows(tpe1, tpe2, renv)
+
       case (x: Type.Var, y: Type.Var) => unifyVars(x.asKinded, y.asKinded, renv)
 
       case (x: Type.Var, _) => (x.kind, tpe2.kind) match {
@@ -101,15 +103,6 @@ object Unification {
       case _ if tpe1.kind == Kind.Effect && tpe2.kind == Kind.Effect =>
         SetUnification.unify(tpe1, tpe2, renv)
 
-      case (row1@Type.Apply(Type.Apply(Type.Cst(TypeConstructor.SchemaRowExtend(_), _), _, _), restRow1, _), row2) =>
-        // Attempt to write the row to match.
-        rewriteSchemaRow(row2, row1, renv) flatMap {
-          case (subst1, restRow2) =>
-            unifyTypes(subst1(restRow1), subst1(restRow2), renv) flatMap {
-              case subst2 => Result.Ok(subst2 @@ subst1)
-            }
-        }
-
       case (Type.Apply(t11, t12, _), Type.Apply(t21, t22, _)) =>
         unifyTypes(t11, t21, renv) match {
           case Result.Ok(subst1) => unifyTypes(subst1(t12), subst1(t22), renv) match {
@@ -121,52 +114,6 @@ object Unification {
 
       case _ => Result.Err(UnificationError.MismatchedTypes(tpe1, tpe2))
     }
-  }
-
-  /**
-    * Attempts to rewrite the given row type `rewrittenRow` such that it shares a first label with `staticRow`.
-    */
-  // TODO: This is a copy of the above function. It would be nice if it could be the same function, but the shape of labels is different.
-  private def rewriteSchemaRow(rewrittenRow: Type, staticRow: Type, renv: RigidityEnv)(implicit flix: Flix): Result[(Substitution, Type), UnificationError] = {
-
-    def visit(row: Type): Result[(Substitution, Type), UnificationError] = (row, staticRow) match {
-      case (Type.Apply(Type.Apply(Type.Cst(TypeConstructor.SchemaRowExtend(label2), _), fieldType2, _), restRow2, loc),
-      Type.Apply(Type.Apply(Type.Cst(TypeConstructor.SchemaRowExtend(label1), _), fieldType1, _), _, _)) =>
-        // Case 1: The row is of the form { label2 :: fieldType2 | restRow2 }
-        if (label1 == label2) {
-          // Case 1.1: The labels match, their types must match.
-          for {
-            subst <- unifyTypes(fieldType1, fieldType2, renv)
-          } yield (subst, restRow2)
-        } else {
-          // Case 1.2: The labels do not match, attempt to match with a label further down.
-          visit(restRow2) map {
-            case (subst, rewrittenRow) => (subst, Type.mkSchemaRowExtend(label2, fieldType2, rewrittenRow, loc))
-          }
-        }
-      case (tvar: Type.Var, Type.Apply(Type.Apply(Type.Cst(TypeConstructor.SchemaRowExtend(label1), _), fieldType1, _), _, _)) =>
-        val tv = tvar.asKinded
-        // Case 2: The row is a type variable.
-        if (staticRow.typeVars contains tv) {
-          Err(UnificationError.OccursCheck(tv, staticRow))
-        } else {
-          // Introduce a fresh type variable to represent one more level of the row.
-          val restRow2 = Type.freshVar(Kind.SchemaRow, tvar.loc)
-          val type2 = Type.mkSchemaRowExtend(label1, fieldType1, restRow2, tvar.loc)
-          val subst = Substitution.singleton(tv.sym, type2)
-          Ok((subst, restRow2))
-        }
-
-      case (Type.Cst(TypeConstructor.SchemaRowEmpty, _), Type.Apply(Type.Apply(Type.Cst(TypeConstructor.SchemaRowExtend(label1), _), fieldType1, _), _, _)) =>
-        // Case 3: The `label` does not exist in the record.
-        Err(UnificationError.UndefinedPredicate(label1, fieldType1, rewrittenRow))
-
-      case _ =>
-        // Case 4: The type is not a row.
-        Err(UnificationError.NonSchemaType(rewrittenRow))
-    }
-
-    visit(rewrittenRow)
   }
 
   /**
